requirements.txt
# ═══════════════════════════════════════════════════════════════════════
# Prism — Unified Python Dependencies
# ═══════════════════════════════════════════════════════════════════════
# pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════════════

# ── Core Framework ─────────────────────────────────────────────────────
fastapi>=0.116,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0

# ── Inter-Service Communication ────────────────────────────────────────
httpx>=0.27,<1.0

# ── Database (SQLite fallback via SQLAlchemy) ──────────────────────────
sqlalchemy>=2.0.23,<3.0

# ── Forecasting (Prophet) ─────────────────────────────────────────────
prophet>=1.1.5,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
python-dateutil>=2.8,<3.0

# ── AI / LLM ──────────────────────────────────────────────────────────
openai>=1.3.6,<3.0

# ── Auth & Validation ─────────────────────────────────────────────────
email-validator>=2.1,<3.0
PyJWT>=2.8,<3.0
bcrypt>=4.0,<5.0
